+++ frozen-collections-core/src/hashers/inline_left_range_hasher.rs
@@ -1,4 +1,4 @@
-use crate::DefaultHashBuilder;
+use crate::DefaultBuildHasher;
 use crate::traits::Hasher;
 use crate::utils::cold;
 use core::hash::{BuildHasher, Hash};
@@ -10,7 +10,7 @@ use alloc::string::String;
 ///
 #[doc = include_str!("../doc_snippets/private_api_warning.md")]
 #[derive(Clone, Debug)]
-pub struct InlineLeftRangeHasher<const RANGE_START: usize, const RANGE_END: usize, BH = DefaultHashBuilder> {
+pub struct InlineLeftRangeHasher<const RANGE_START: usize, const RANGE_END: usize, BH = DefaultBuildHasher> {
     bh: BH,
 }
 
@@ -78,7 +78,7 @@ mod tests {
 
     #[test]
     fn test_left_range_hasher_hash_slice() {
-        let hasher = InlineLeftRangeHasher::<0, 3>::new(DefaultHashBuilder::default());
+        let hasher = InlineLeftRangeHasher::<0, 3>::new(DefaultBuildHasher::default());
         assert_eq!(
             hasher.hash_one(vec![1, 2, 3, 4].as_slice()),
             hasher.bh.hash_one(vec![1, 2, 3].as_slice())
@@ -88,21 +88,21 @@ mod tests {
 
     #[test]
     fn test_left_range_hasher_hash_string() {
-        let hasher = InlineLeftRangeHasher::<0, 3>::new(DefaultHashBuilder::default());
+        let hasher = InlineLeftRangeHasher::<0, 3>::new(DefaultBuildHasher::default());
         assert_eq!(hasher.hash_one(&"abcd".to_string()), hasher.bh.hash_one(b"abc"));
         assert_eq!(hasher.hash_one(&"ab".to_string()), 0);
     }
 
     #[test]
     fn test_left_range_hasher_hash_str_ref() {
-        let hasher = InlineLeftRangeHasher::<0, 3>::new(DefaultHashBuilder::default());
+        let hasher = InlineLeftRangeHasher::<0, 3>::new(DefaultBuildHasher::default());
         assert_eq!(hasher.hash_one(&"abcd"), hasher.bh.hash_one(b"abc"));
         assert_eq!(hasher.hash_one(&"ab"), 0);
     }
 
     #[test]
     fn test_left_range_hasher_hash_str() {
-        let hasher = InlineLeftRangeHasher::<0, 3>::new(DefaultHashBuilder::default());
+        let hasher = InlineLeftRangeHasher::<0, 3>::new(DefaultBuildHasher::default());
         assert_eq!(hasher.hash_one("abcd"), hasher.bh.hash_one(b"abc"));
         assert_eq!(hasher.hash_one("ab"), 0);
     }