+++ frozen-collections-core/src/hashers/inline_right_range_hasher.rs
@@ -1,4 +1,4 @@
-use crate::DefaultHashBuilder;
+use crate::DefaultBuildHasher;
 use crate::traits::Hasher;
 use crate::utils::cold;
 use core::hash::{BuildHasher, Hash};
@@ -10,7 +10,7 @@ use alloc::string::String;
 ///
 #[doc = include_str!("../doc_snippets/private_api_warning.md")]
 #[derive(Clone, Debug)]
-pub struct InlineRightRangeHasher<const RANGE_START: usize, const RANGE_END: usize, BH = DefaultHashBuilder> {
+pub struct InlineRightRangeHasher<const RANGE_START: usize, const RANGE_END: usize, BH = DefaultBuildHasher> {
     bh: BH,
 }
 
@@ -81,7 +81,7 @@ mod tests {
 
     #[test]
     fn test_right_range_hasher_hash_slice() {
-        let hasher = InlineRightRangeHasher::<1, 3>::new(DefaultHashBuilder::default());
+        let hasher = InlineRightRangeHasher::<1, 3>::new(DefaultBuildHasher::default());
         assert_eq!(
             hasher.hash_one(vec![1, 2, 3, 4].as_slice()),
             hasher.bh.hash_one(vec![2, 3].as_slice())
@@ -95,7 +95,7 @@ mod tests {
 
     #[test]
     fn test_right_range_hasher_hash_string() {
-        let hasher = InlineRightRangeHasher::<3, 5>::new(DefaultHashBuilder::default());
+        let hasher = InlineRightRangeHasher::<3, 5>::new(DefaultBuildHasher::default());
         assert_eq!(hasher.hash_one(&"abcdef".to_string()), hasher.bh.hash_one(b"bc"));
         assert_eq!(hasher.hash_one(&"abcdefghijklmn".to_string()), hasher.bh.hash_one(b"jk"));
         assert_eq!(hasher.hash_one(&"a".to_string()), 0);
@@ -103,7 +103,7 @@ mod tests {
 
     #[test]
     fn test_right_range_hasher_hash_str_ref() {
-        let hasher = InlineRightRangeHasher::<1, 3>::new(DefaultHashBuilder::default());
+        let hasher = InlineRightRangeHasher::<1, 3>::new(DefaultBuildHasher::default());
         assert_eq!(hasher.hash_one(&"abcd"), hasher.bh.hash_one(b"bc"));
         assert_eq!(hasher.hash_one(&"abcdefghijklmn"), hasher.bh.hash_one(b"lm"));
         assert_eq!(hasher.hash_one(&"a"), 0);
@@ -111,7 +111,7 @@ mod tests {
 
     #[test]
     fn test_right_range_hasher_hash_str() {
-        let hasher = InlineRightRangeHasher::<1, 3>::new(DefaultHashBuilder::default());
+        let hasher = InlineRightRangeHasher::<1, 3>::new(DefaultBuildHasher::default());
         assert_eq!(hasher.hash_one("abcd"), hasher.bh.hash_one(b"bc"));
         assert_eq!(hasher.hash_one("abcdefghijklmn"), hasher.bh.hash_one(b"lm"));
         assert_eq!(hasher.hash_one("a"), 0);