+++ frozen-collections-core/src/lib.rs
@@ -34,8 +34,8 @@ mod emit;
 
 /// The default hash builder used by the frozen collections.
 #[cfg(test)]
-pub type DefaultHashBuilder = foldhash::fast::FixedState;
+pub type DefaultBuildHasher = foldhash::fast::FixedState;
 
 /// The default hash builder used by the frozen collections.
 #[cfg(not(test))]
-pub type DefaultHashBuilder = foldhash::fast::RandomState;
+pub type DefaultBuildHasher = foldhash::fast::RandomState;